more location refactoring
