+++ cli/output/jsonw/metadatawriter.go
@@ -3,11 +3,12 @@ package jsonw
 import (
 	"io"
 
+	"github.com/neilotoole/sq/libsq/source/location"
+
 	"github.com/samber/lo"
 
 	"github.com/neilotoole/sq/cli/output"
 	"github.com/neilotoole/sq/libsq/driver"
-	"github.com/neilotoole/sq/libsq/source"
 	"github.com/neilotoole/sq/libsq/source/metadata"
 )
 
@@ -36,7 +37,7 @@ func (w *mdWriter) TableMetadata(md *metadata.Table) error {
 // SourceMetadata implements output.MetadataWriter.
 func (w *mdWriter) SourceMetadata(md *metadata.Source, showSchema bool) error {
 	md2 := *md // Shallow copy is fine
-	md2.Location = source.RedactLocation(md2.Location)
+	md2.Location = location.RedactLocation(md2.Location)
 
 	if showSchema {
 		return writeJSON(w.out, w.pr, &md2)
