+++ cli/output/tablew/metadatawriter.go
@@ -9,6 +9,8 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/neilotoole/sq/libsq/source/location"
+
 	"github.com/samber/lo"
 
 	"github.com/neilotoole/sq/cli/output"
@@ -16,7 +18,6 @@ import (
 	"github.com/neilotoole/sq/libsq/core/kind"
 	"github.com/neilotoole/sq/libsq/core/stringz"
 	"github.com/neilotoole/sq/libsq/driver"
-	"github.com/neilotoole/sq/libsq/source"
 	"github.com/neilotoole/sq/libsq/source/metadata"
 )
 
@@ -124,7 +125,7 @@ func (w *mdWriter) doSourceMetaNoSchema(md *metadata.Source) error {
 		md.Name,
 		md.FQName,
 		w.tbl.pr.Number.Sprint(stringz.ByteSized(md.Size, 1, "")),
-		source.RedactLocation(md.Location),
+		location.RedactLocation(md.Location),
 	}
 
 	w.tbl.tblImpl.SetHeader(headers)
@@ -257,7 +258,7 @@ func (w *mdWriter) doSourceMetaFull(md *metadata.Source) error {
 		w.tbl.pr.Number.Sprint(stringz.ByteSized(md.Size, 1, "")),
 		fmt.Sprintf("%d", md.TableCount),
 		fmt.Sprintf("%d", md.ViewCount),
-		source.RedactLocation(md.Location),
+		location.RedactLocation(md.Location),
 	}
 
 	w.tbl.tblImpl.SetHeader(headers)
