+++ libsq/source/internal_test.go
@@ -1,10 +1,7 @@
 package source
 
 import (
-	"runtime"
 	"testing"
-
-	"github.com/neilotoole/sq/libsq/source/location"
 
 	"github.com/stretchr/testify/require"
 
@@ -15,150 +12,6 @@ import (
 var (
 	GroupsFilterOnlyDirectChildren = groupsFilterOnlyDirectChildren
 )
-
-func TestParseLoc(t *testing.T) {
-	const (
-		dbuser = "sakila"
-		dbpass = "p_ssW0rd"
-	)
-
-	testCases := []struct {
-		loc     string
-		want    location.ParsedLoc
-		wantErr bool
-		windows bool
-	}{
-		{
-			loc:  "/path/to/sakila.xlsx",
-			want: location.ParsedLoc{Name: "sakila", Ext: ".xlsx"},
-		},
-		{
-			loc:  "relative/path/to/sakila.xlsx",
-			want: location.ParsedLoc{Name: "sakila", Ext: ".xlsx"},
-		},
-		{
-			loc:  "./relative/path/to/sakila.xlsx",
-			want: location.ParsedLoc{Name: "sakila", Ext: ".xlsx"},
-		},
-		{
-			loc:  "https://server:8080/path/to/sakila.xlsx",
-			want: location.ParsedLoc{Scheme: "https", Hostname: "server", Port: 8080, Name: "sakila", Ext: ".xlsx"},
-		},
-		{
-			loc:  "http://server/path/to/sakila.xlsx?param=val&param2=val2",
-			want: location.ParsedLoc{Scheme: "http", Hostname: "server", Name: "sakila", Ext: ".xlsx"},
-		},
-		{
-			loc:     "sqlite3:/path/to/sakila.db",
-			wantErr: true,
-		}, // the scheme is malformed (should be "sqlite3://...")
-		{
-			loc: "sqlite3:///path/to/sakila.sqlite",
-			want: location.ParsedLoc{
-				DriverType: typeSL3, Scheme: "sqlite3", Name: "sakila", Ext: ".sqlite",
-				DSN: "/path/to/sakila.sqlite",
-			},
-		},
-		{
-			loc:     `sqlite3://C:\path\to\sakila.sqlite`,
-			windows: true,
-			want: location.ParsedLoc{
-				DriverType: typeSL3, Scheme: "sqlite3", Name: "sakila", Ext: ".sqlite",
-				DSN: `C:\path\to\sakila.sqlite`,
-			},
-		},
-		{
-			loc:     `sqlite3://C:\path\to\sakila.sqlite?param=val`,
-			windows: true,
-			want: location.ParsedLoc{
-				DriverType: typeSL3, Scheme: "sqlite3", Name: "sakila", Ext: ".sqlite",
-				DSN: `C:\path\to\sakila.sqlite?param=val`,
-			},
-		},
-		{
-			loc: "sqlite3:///path/to/sakila",
-			want: location.ParsedLoc{
-				DriverType: typeSL3, Scheme: "sqlite3", Name: "sakila", DSN: "/path/to/sakila",
-			},
-		},
-		{
-			loc: "sqlite3://path/to/sakila.db",
-			want: location.ParsedLoc{
-				DriverType: typeSL3, Scheme: "sqlite3", Name: "sakila", Ext: ".db", DSN: "path/to/sakila.db",
-			},
-		},
-		{
-			loc: "sqlite3:///path/to/sakila.db",
-			want: location.ParsedLoc{
-				DriverType: typeSL3, Scheme: "sqlite3", Name: "sakila", Ext: ".db", DSN: "/path/to/sakila.db",
-			},
-		},
-		{
-			loc: "sqlserver://sakila:p_ssW0rd@localhost?database=sakila",
-			want: location.ParsedLoc{
-				DriverType: typeMS, Scheme: "sqlserver", User: dbuser, Pass: dbpass, Hostname: "localhost",
-				Name: "sakila", DSN: "sqlserver://sakila:p_ssW0rd@localhost?database=sakila",
-			},
-		},
-		{
-			loc: "sqlserver://sakila:p_ssW0rd@server:1433?database=sakila",
-			want: location.ParsedLoc{
-				DriverType: typeMS, Scheme: "sqlserver", User: dbuser, Pass: dbpass, Hostname: "server",
-				Port: 1433, Name: "sakila",
-				DSN: "sqlserver://sakila:p_ssW0rd@server:1433?database=sakila",
-			},
-		},
-		{
-			loc: "postgres://sakila:p_ssW0rd@localhost/sakila?sslmode=disable",
-			want: location.ParsedLoc{
-				DriverType: typePg, Scheme: "postgres", User: dbuser, Pass: dbpass, Hostname: "localhost",
-				Name: "sakila", DSN: "dbname=sakila host=localhost password=p_ssW0rd sslmode=disable user=sakila",
-			},
-		},
-		{
-			loc: "postgres://sakila:p_ssW0rd@server:5432/sakila?sslmode=disable",
-			want: location.ParsedLoc{
-				DriverType: typePg, Scheme: "postgres", User: dbuser, Pass: dbpass, Hostname: "server", Port: 5432,
-				Name: "sakila",
-				DSN:  "dbname=sakila host=server password=p_ssW0rd port=5432 sslmode=disable user=sakila",
-			},
-		},
-		{
-			loc: "mysql://sakila:p_ssW0rd@localhost/sakila",
-			want: location.ParsedLoc{
-				DriverType: typeMy, Scheme: "mysql", User: dbuser, Pass: dbpass, Hostname: "localhost",
-				Name: "sakila", DSN: "sakila:p_ssW0rd@tcp(localhost:3306)/sakila",
-			},
-		},
-		{
-			loc: "mysql://sakila:p_ssW0rd@server:3306/sakila",
-			want: location.ParsedLoc{
-				DriverType: typeMy, Scheme: "mysql", User: dbuser, Pass: dbpass, Hostname: "server", Port: 3306,
-				Name: "sakila", DSN: "sakila:p_ssW0rd@tcp(server:3306)/sakila",
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tu.Name(1, RedactLocation(tc.loc)), func(t *testing.T) {
-			if tc.windows && runtime.GOOS != "windows" {
-				return
-			}
-
-			tc.want.Loc = tc.loc // set this here rather than verbosely in the setup
-			got, gotErr := location.ParseLocation(tc.loc)
-			if tc.wantErr {
-				require.Error(t, gotErr)
-				require.Nil(t, got)
-				return
-			}
-
-			require.NoError(t, gotErr)
-			require.Equal(t, tc.want, *got)
-		})
-	}
-}
 
 func TestGroupsFilterOnlyDirectChildren(t *testing.T) {
 	testCases := []struct {
